check if there is a .pdf in the folder

+++ parsetickets.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 # go to tickets folders, convert *.pdf to .txt and delete *.pdf
 cd ~/consum/consum_project/data/tickets_pdf
-for file in *.pdf; do
-  base=${file%.*}
-  pdftotext -layout "$file" "$base.txt"
-  rm $file
-done;
+if [ -e *.pdf ]; then
+  for file in *.pdf; do
+    base=${file%.*}
+    pdftotext -layout "$file" "$base.txt"
+    rm $file
+  done;
+fi
